RAG_Project/pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "pypdf>=6.9.2",
    "sentence-transformers>=5.3.0",
    "tiktoken>=0.12.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
    "streamlit>=1.30.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "langchain-classic>=1.0.3",
    "langchain-huggingface>=1.2.2",
    "langchain-chroma>=1.1.0",
    "chromadb>=1.5.5",
    "langchain-core==1.2.31",
]
